[Refactor] Extract split scan morsel classes
