+++ be/src/connector/data_source.h
@@ -22,6 +22,7 @@
 #include "common/runtime_profile.h"
 #include "common/statusor.h"
 #include "exec/pipeline/scan/morsel.h"
+#include "exec/pipeline/scan/scan_morsel.h"
 #include "exec/runtime_filter/runtime_filter_probe.h"
 #include "gen_cpp/InternalService_types.h"
 #include "gen_cpp/PlanNodes_types.h"
