+++ be/src/exec/CMakeLists.txt
@@ -238,6 +238,8 @@ set(PIPELINE_OPERATORS_BATCH_SRCS
     pipeline/scan/balanced_chunk_buffer.cpp
     pipeline/scan/chunk_source.cpp
     pipeline/scan/morsel.cpp
+    pipeline/scan/scan_morsel.cpp
+    pipeline/scan/split_scan_morsel.cpp
     pipeline/scan/scan_operator.cpp
     pipeline/scan/olap_chunk_source.cpp
     pipeline/scan/olap_scan_operator.cpp
