+++ be/src/exec/pipeline/scan/connector_scan_operator.cpp
@@ -21,6 +21,7 @@
 #include "exec/pipeline/pipeline_driver.h"
 #include "exec/pipeline/query_context.h"
 #include "exec/pipeline/scan/balanced_chunk_buffer.h"
+#include "exec/pipeline/scan/scan_morsel.h"
 #include "exprs/expr_executor.h"
 #include "runtime/descriptors.h"
 #include "runtime/exec_env.h"
