+++ be/src/exec/pipeline/scan/meta_chunk_source.cpp
@@ -15,6 +15,7 @@
 #include "exec/pipeline/scan/meta_chunk_source.h"
 
 #include "exec/pipeline/scan/meta_scan_operator.h"
+#include "exec/pipeline/scan/scan_morsel.h"
 #include "exec/workgroup/work_group.h"
 
 namespace starrocks::pipeline {
