+++ be/src/exec/pipeline/scan/morsel.h
@@ -15,11 +15,16 @@
 #pragma once
 
 #include <atomic>
+#include <deque>
+#include <map>
+#include <memory>
 #include <mutex>
-#include <optional>
+#include <string>
+#include <unordered_map>
+#include <vector>
 
+#include "exec/pipeline/scan/scan_morsel.h"
 #include "exec/query_cache/ticket_checker.h"
-#include "gen_cpp/InternalService_types.h"
 #include "runtime/mem_pool.h"
 #include "storage/olap_common.h"
 #include "storage/olap_tuple.h"
@@ -34,28 +39,17 @@ namespace starrocks {
 struct OlapScanRange;
 class BaseTablet;
 using BaseTabletSharedPtr = std::shared_ptr<BaseTablet>;
-class BaseRowset;
-using BaseRowsetSharedPtr = std::shared_ptr<BaseRowset>;
 class Segment;
 using SegmentSharedPtr = std::shared_ptr<Segment>;
 
 struct TabletReaderParams;
 class SeekTuple;
-struct RowidRangeOption;
-using RowidRangeOptionPtr = std::shared_ptr<RowidRangeOption>;
-struct ShortKeyRangeOption;
-using ShortKeyRangeOptionPtr = std::shared_ptr<ShortKeyRangeOption>;
 struct ShortKeyOption;
 using ShortKeyOptionPtr = std::unique_ptr<ShortKeyOption>;
 class Schema;
 using SchemaPtr = std::shared_ptr<Schema>;
 
 namespace pipeline {
-
-class ScanMorsel;
-using Morsel = ScanMorsel;
-using MorselPtr = std::unique_ptr<Morsel>;
-using Morsels = std::vector<MorselPtr>;
 
 class MorselQueue;
 class SplitMorselQueue;
@@ -64,180 +58,6 @@ using MorselQueueMap = std::unordered_map<int32_t, MorselQueuePtr>;
 class MorselQueueFactory;
 using MorselQueueFactoryPtr = std::unique_ptr<MorselQueueFactory>;
 using MorselQueueFactoryMap = std::unordered_map<int32_t, MorselQueueFactoryPtr>;
-
-/// Morsel.
-class ScanMorselX {
-public:
-    explicit ScanMorselX(int32_t plan_node_id) : _plan_node_id(plan_node_id) {}
-    virtual ~ScanMorselX() = default;
-
-    int32_t get_plan_node_id() const { return _plan_node_id; }
-
-    virtual void init_tablet_reader_params(TabletReaderParams* params) {}
-
-    virtual std::tuple<int64_t, int64_t> get_lane_owner_and_version() const {
-        return std::tuple<int64_t, int64_t>{0L, 0L};
-    }
-
-    // from_version is used when reading incremental rowsets. in default, from_version = 0 means all of the rowsets
-    // will be read out. In multi-version cache mechanism, when probing the cache and finding that cached result has
-    // stale version, then incremental rowsets in the version range from the cached version till required version
-    // should be read out and merged with the cache result, here from_version is cached version.
-    void set_from_version(int64_t from_version) { _from_version = from_version; }
-    int64_t from_version() { return _from_version; }
-
-    void set_rowsets(const std::vector<BaseRowsetSharedPtr>& rowsets) { _rowsets = &rowsets; }
-    void set_delta_rowsets(std::vector<BaseRowsetSharedPtr>&& delta_rowsets) {
-        _delta_rowsets = std::move(delta_rowsets);
-    }
-    const std::vector<BaseRowsetSharedPtr>& rowsets() const {
-        if (_delta_rowsets.has_value()) {
-            return _delta_rowsets.value();
-        } else {
-            return *_rowsets;
-        }
-    }
-
-    virtual const std::unordered_set<std::string>& skip_min_max_metrics() const {
-        static const std::unordered_set<std::string> metrics;
-        return metrics;
-    }
-
-private:
-    int32_t _plan_node_id;
-    int64_t _from_version = 0;
-
-    static const std::vector<BaseRowsetSharedPtr> kEmptyRowsets;
-    // _rowsets is owned by MorselQueue, whose lifecycle is longer than that of Morsel.
-    const std::vector<BaseRowsetSharedPtr>* _rowsets = &kEmptyRowsets;
-    std::optional<std::vector<BaseRowsetSharedPtr>> _delta_rowsets;
-};
-
-class ScanSplitContext {
-public:
-    virtual ~ScanSplitContext() = default;
-    void set_last_split(bool v) { _is_last_split = v; }
-    bool is_last_split() const { return _is_last_split; }
-
-private:
-    bool _is_last_split = false;
-};
-
-struct LakeSplitContext : public ScanSplitContext {
-    // physical split
-    RowidRangeOptionPtr rowid_range;
-    // logical split
-    ShortKeyRangesOptionPtr short_key_range;
-    std::shared_ptr<SplitMorselQueue> split_morsel_queue = nullptr;
-};
-
-using ScanSplitContextPtr = std::unique_ptr<ScanSplitContext>;
-
-class ScanMorsel : public ScanMorselX {
-public:
-    ScanMorsel(int32_t plan_node_id, const TScanRange& scan_range)
-            : ScanMorselX(plan_node_id), _scan_range(std::make_unique<TScanRange>(scan_range)) {
-        if (_scan_range->__isset.internal_scan_range) {
-            _owner_id = _scan_range->internal_scan_range.tablet_id;
-            auto str_version = _scan_range->internal_scan_range.version;
-            _version = strtol(str_version.c_str(), nullptr, 10);
-            _owner_id = _scan_range->internal_scan_range.__isset.bucket_sequence
-                                ? _scan_range->internal_scan_range.bucket_sequence
-                                : _owner_id;
-            _partition_id = _scan_range->internal_scan_range.partition_id;
-            _has_owner_id = true;
-        }
-        if (_scan_range->__isset.binlog_scan_range) {
-            _owner_id = _scan_range->binlog_scan_range.tablet_id;
-            _has_owner_id = true;
-        }
-    }
-
-    ScanMorsel(int32_t plan_node_id, const TScanRangeParams& scan_range)
-            : ScanMorsel(plan_node_id, scan_range.scan_range) {}
-
-    ~ScanMorsel() override = default;
-
-    TScanRange* get_scan_range() { return _scan_range.get(); }
-
-    TInternalScanRange* get_olap_scan_range() { return &(_scan_range->internal_scan_range); }
-
-    std::tuple<int64_t, int64_t> get_lane_owner_and_version() const override {
-        return std::tuple<int64_t, int64_t>{_owner_id, _version};
-    }
-
-    void set_split_context(ScanSplitContextPtr&& split_context) {
-        if (split_context == nullptr) return;
-        _split_context = std::move(split_context);
-        _is_last_split = _split_context->is_last_split();
-    }
-    ScanSplitContext* get_split_context() { return _split_context.get(); }
-
-    bool has_owner_id() const { return _has_owner_id; }
-    int32_t owner_id() const { return _owner_id; }
-    int32_t partition_id() const { return _partition_id; }
-
-    bool is_last_split() const { return _is_last_split; }
-    void set_last_split(bool v) { _is_last_split = v; }
-
-    bool is_ticket_checker_entered() const { return _ticket_checker_entered; }
-    void set_ticket_checker_entered(bool v) { _ticket_checker_entered = v; }
-
-    static void build_scan_morsels(int node_id, const std::vector<TScanRangeParams>& scan_ranges,
-                                   bool accept_empty_scan_ranges, pipeline::Morsels* morsels, bool* has_more_morsel);
-    static bool has_more_scan_ranges(const std::vector<TScanRangeParams>& scan_ranges);
-
-private:
-    std::unique_ptr<TScanRange> _scan_range;
-    ScanSplitContextPtr _split_context = nullptr;
-    bool _has_owner_id = false;
-    int64_t _owner_id = 0;
-    int64_t _version = 0;
-    int64_t _partition_id = 0;
-    bool _is_last_split = true;
-    bool _ticket_checker_entered = false;
-};
-
-class PhysicalSplitScanMorsel final : public ScanMorsel {
-public:
-    PhysicalSplitScanMorsel(int32_t plan_node_id, const TScanRange& scan_range, RowidRangeOptionPtr rowid_range_option)
-            : ScanMorsel(plan_node_id, scan_range), _rowid_range_option(std::move(rowid_range_option)) {}
-
-    ~PhysicalSplitScanMorsel() override = default;
-
-    void init_tablet_reader_params(TabletReaderParams* params) override;
-
-    const std::unordered_set<std::string>& skip_min_max_metrics() const override {
-        static const std::unordered_set<std::string> metrics{"ShortKeyFilterRows", "SegmentZoneMapFilterRows"};
-        return metrics;
-    }
-
-    RowidRangeOptionPtr get_rowid_range_option() { return _rowid_range_option; }
-
-private:
-    RowidRangeOptionPtr _rowid_range_option;
-};
-
-class LogicalSplitScanMorsel final : public ScanMorsel {
-public:
-    LogicalSplitScanMorsel(int32_t plan_node_id, const TScanRange& scan_range,
-                           ShortKeyRangesOptionPtr short_key_ranges_option)
-            : ScanMorsel(plan_node_id, scan_range), _short_key_ranges_option(std::move(short_key_ranges_option)) {}
-
-    ~LogicalSplitScanMorsel() override = default;
-
-    void init_tablet_reader_params(TabletReaderParams* params) override;
-
-    const std::unordered_set<std::string>& skip_min_max_metrics() const override {
-        static const std::unordered_set<std::string> metrics{"ShortKeyFilterRows", "SegmentZoneMapFilterRows"};
-        return metrics;
-    }
-
-    ShortKeyRangesOptionPtr get_short_key_ranges_option() { return _short_key_ranges_option; }
-
-private:
-    ShortKeyRangesOptionPtr _short_key_ranges_option;
-};
 
 /// MorselQueueFactory.
 class MorselQueueFactory {
