+++ be/src/exec/pipeline/source_operator.h
@@ -19,6 +19,7 @@
 #include "exec/pipeline/adaptive/adaptive_fwd.h"
 #include "exec/pipeline/operator_factory.h"
 #include "exec/pipeline/scan/chunk_source.h"
+#include "exec/pipeline/scan/morsel.h"
 #include "exec/pipeline/schedule/observer.h"
 #include "exec/workgroup/work_group_fwd.h"
 #include "gen_cpp/Partitions_types.h"
