+++ be/src/exec/scan_node.cpp
@@ -37,6 +37,7 @@
 #include "common/config_scan_io_fwd.h"
 #include "exec/pipeline/query_context.h"
 #include "exec/pipeline/scan/morsel.h"
+#include "exec/pipeline/scan/scan_morsel.h"
 #include "exprs/expr_factory.h"
 #include "runtime/runtime_state.h"
